requirements.txt
# Final AI Service Requirements - Production Ready

# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI/ML Core
torch>=2.0.0
transformers>=4.35.0

# Image Processing
Pillow>=10.0.0

# Specialized Models (High Accuracy)
nudenet>=3.0.0  # High-accuracy NSFW detection

# Text Moderation Models
# martin-ha/toxic-comment-model - Downloaded automatically by transformers

# Document Processing Libraries
PyPDF2>=3.0.0  # PDF text extraction
python-docx>=0.8.11  # DOCX text extraction
openpyxl>=3.1.0  # Excel file processing
python-pptx>=0.6.21  # PowerPoint file processing
pdf2image>=1.16.0  # PDF to image conversion
PyMuPDF>=1.23.0  # Advanced PDF processing and image extraction
# Note: python-magic removed due to Windows compatibility issues
# Using built-in mimetypes and magic bytes detection instead

# Text Processing & Hashing
datasketch>=1.5.0
simhash>=2.1.0

# Utilities
numpy>=1.24.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
